move landingpage ad into assets
