+++ src/components/LandingPage/LandingPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Box from '@material-ui/core/Box';
 import Banner from './Banner';
 import Info from './Info';
-import Ad from './Advertisement';
+import Ad from '../assets/Omega_ad-Banner';
 import InstructionsPage from '../InstructionsPage/InstructionsPage';
 
 const LandingPage = () => {
